refactor: rename keydown listener
